Move mocks to a separate file
